Updated routing table implementation for  -translate the IDT_DSF_RT_ values to the IDT_RXS_DSF_RT_ values -check the RXS_RIO_SPX_ERR_STAT_PORT_UNAVL bit to see if the port is available or not -use IDT_DSF_rt_alloc_mc_mask and IDT_DSF_rt_dealloc_mc_mask for RXS -search the rt->dom_table for the MC mask value -use the default routine for rioSetAddrMode -Bind_rxs2448_DSF_support/bind_rxs1632_dsf_support collapsed into a single routine -DARDB_rioDeviceSupportedDefault enhanced to support the new register blocks

+++ fabric_management/librio_switch/src/IDT_CPS_RT_Common_API.c
@@ -460,7 +460,7 @@ uint32_t read_rte_entries( DAR_DEV_INFO_t            *dev_info,
    //
    for (destID = 0; destID < IDT_DAR_RT_DEV_TABLE_SIZE; destID++)
    {
-      uint8_t mask_idx;
+      uint32_t mask_idx;
 
       rt->dev_table[destID].changed = false;
